fix: stop canonical inheritance, enrich metadata and add breadcrumb json-ld
