+++ src/app/[locale]/(auth)/sign-in/page.tsx
@@ -20,6 +20,7 @@ export async function generateMetadata({
 
   return {
     title: t("auth.signIn.metadata.title"),
+    robots: { index: false },
   };
 }
