+++ src/app/[locale]/(business)/(with-header)/friend-requests/accept/page.tsx
@@ -14,6 +14,12 @@ import { Link, redirect } from "@/lib/i18n";
 import { Routes } from "@/lib/routes";
 import { generatePath } from "@/lib/routes/utils";
 
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  robots: { index: false },
+};
+
 const AcceptFriendRequestPage = async ({
   params,
   searchParams,
