+++ src/app/[locale]/(business)/(with-header)/friend-requests/deny/page.tsx
@@ -14,6 +14,12 @@ import { redirect } from "@/lib/i18n";
 import { Routes } from "@/lib/routes";
 import { generatePath } from "@/lib/routes/utils";
 
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  robots: { index: false },
+};
+
 const RejectFriendRequestPage = async ({
   params,
   searchParams,
