+++ src/app/[locale]/(business)/(with-header)/settings/page.tsx
@@ -15,6 +15,7 @@ export async function generateMetadata({
 
   return {
     title: t("settingsPage.metadata.title"),
+    robots: { index: false },
   };
 }
