+++ src/lib/seo/index.ts
@@ -1,6 +1,8 @@
 import { publicConfig } from "@/lib/config/client-config";
 import { routing } from "@/lib/i18n";
 
+import type { BreadcrumbList, WithContext } from "schema-dts";
+
 export const URLS_PER_SITEMAP = 50000;
 
 export const getAbsoluteUrl = (path: string) => {
@@ -16,6 +18,19 @@ export const getAbsoluteUrl = (path: string) => {
 
   return `${baseUrl}/${path}`;
 };
+
+export const getBreadcrumbListJsonLd = (
+  items: Array<{ name: string; path: string }>,
+): WithContext<BreadcrumbList> => ({
+  "@context": "https://schema.org",
+  "@type": "BreadcrumbList",
+  itemListElement: items.map((item, index) => ({
+    "@type": "ListItem",
+    position: index + 1,
+    name: item.name,
+    item: getAbsoluteUrl(item.path),
+  })),
+});
 
 export const getAlternates = (path: string, locale?: string) => {
   const normalizedPath = path === "/" ? "" : path;
